Move stuff around + bit of readme
